+++ projects/rocfft/clients/tests/gtest_main.cpp
@@ -451,9 +451,9 @@ int main(int argc, char* argv[])
             n_random_tests = 10;
         });
 
-    app.add_flag("--callback", "Inject load/store callbacks")->each([&](const std::string&) {
-        manual_params.run_callbacks = true;
-    });
+    app.add_flag(
+           "--callback", manual_params.run_callbacks, "Inject load/store callbacks: none, funcptr")
+        ->default_val("none");
 
     app.add_option("--seed", random_seed, "Random seed; if unset, use an actual random seed")
         ->default_val(default_seed_dev());