+++ projects/rocfft/shared/accuracy_test.h
@@ -135,16 +135,16 @@ inline void execute_gpu_fft(Tparams&              params,
     std::vector<void*> store_cb_func;
     std::vector<void*> store_cb_data;
 
-    // Legacy callbacks are provided at execution time
-    if(params.run_callbacks)
+    // Function pointer callbacks are provided at execution time
+    if(params.run_callbacks == fft_callback_type_funcptr)
     {
-        get_rank_load_callbacks(
+        get_rank_load_callbacks_funcptr(
             params, load_cb_func, load_cb_data, round_trip_inverse, all_cb_data);
-        get_rank_store_callbacks(
+        get_rank_store_callbacks_funcptr(
             params, store_cb_func, store_cb_data, round_trip_inverse, all_cb_data);
 
-        auto fft_status
-            = params.set_callbacks(&load_cb_func, &load_cb_data, &store_cb_func, &store_cb_data);
+        auto fft_status = params.set_funcptr_callbacks(
+            &load_cb_func, &load_cb_data, &store_cb_func, &store_cb_data);
         if(fft_status != fft_status_success)
             throw std::runtime_error("set callback failure");
     }